This is an attempt (work around) to restore Blob quest area on map without tainting UI while opening map in combat. Here what it does: - it hide auto the map if we enter in combat. - if we enter in combat, quest tracking POI/Blob is disabled. - if we exit combat, quest tracking POI/blob is restored.
Finally, just note that this is not yet tested in game. Fork this change at your risk. Let's me know if you did a test of this change and if taint are gone. This is I think the best we can do to avoid taint with POI/BLOB on map.